+++ cmd/manager/main.go
@@ -123,7 +123,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	log.Info("Registering Components.", "")
+	log.Info("Registering Components.")
 
 	// Setup Scheme for all resources
 	if err := apis.AddToScheme(mgr.GetScheme()); err != nil {