+++ pkg/config/config.go
@@ -8,14 +8,6 @@ type Config struct {
 	MDCContainerName        string
 	OauthProxyContainerName string
 
-	MDCImageStreamName        string
-	MDCImageStreamTag         string
-	OauthProxyImageStreamName string
-	OauthProxyImageStreamTag  string
-
-	MDCImageStreamInitialImage        string
-	OauthProxyImageStreamInitialImage string
-
 	UnifiedPushDocumentationURL        string
 	IdentityManagementDocumentationURL string
 	DataSyncDocumentationURL           string
@@ -28,15 +20,6 @@ func New() Config {
 
 		MDCContainerName:        getEnv("MDC_CONTAINER_NAME", "mdc"),
 		OauthProxyContainerName: getEnv("OAUTH_PROXY_CONTAINER_NAME", "mdc-oauth-proxy"),
-
-		MDCImageStreamName:        getEnv("MDC_IMAGE_STREAM_NAME", "mdc-imagestream"),
-		MDCImageStreamTag:         getEnv("MDC_IMAGE_STREAM_TAG", "master"),
-		OauthProxyImageStreamName: getEnv("OAUTH_PROXY_IMAGE_STREAM_NAME", "mdc-oauth-proxy-imagestream"),
-		OauthProxyImageStreamTag:  getEnv("OAUTH_PROXY_IMAGE_STREAM_TAG", "latest"),
-
-		// these are used when the image stream does not exist and created for the first time by the operator
-		MDCImageStreamInitialImage:        getEnv("MDC_IMAGE_STREAM_INITIAL_IMAGE", "quay.io/aerogear/mobile-developer-console:master"),
-		OauthProxyImageStreamInitialImage: getEnv("OAUTH_PROXY_IMAGE_STREAM_INITIAL_IMAGE", "quay.io/openshift/origin-oauth-proxy:4.2.0"),
 
 		// override the default links displayed in MDC for each of the mobile services
 		UnifiedPushDocumentationURL:        getEnv("UPS_DOCUMENTATION_URL", "https://docs.aerogear.org/limited-availability/upstream/ups.html"),